L'outil make-finance-data génère les agrégations
